Split out autodiff_gradient.h from autodiff.h
The premise is that autodiff.h gets autodiff code, that gradient.h gets
gradient code, and that autodiff_gradient.h gets the go-between code.

Specifically:
- Move autoDiffToGradientMatrix into autodiff_gradient.h
- Move the type-helper structs into internal namespaces (and separate files).
- Add file overview comments.
- Update the includes in downstream code.
