+++ drake/math/autodiff_gradient.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// autodiff_gradient.h is a stand-alone header.
+
+#include "drake/math/autodiff_gradient.h"
