+++ drake/math/test/autodiff_test.cc
@@ -6,6 +6,7 @@
 #include "gtest/gtest.h"
 
 #include "drake/common/eigen_types.h"
+#include "drake/math/autodiff_gradient.h"
 #include "drake/util/eigen_matrix_compare.h"
 
 using Eigen::MatrixXd;
